Minor Changes to Format & Error Handling
Updated 'hovertext' labels to include units and date/time info.

Added a condition to only draw red 'out of control' points when there are actually rule violations (this was causing an error).

Updated the traffic lights to ACTUALLY use data from the past 24hrs (previously it was from midnight the day before the day of access)